requirements.txt
tkinter 
pillow 
speechrecognition 
sounddevice 
numpy 
scipy 
